Gitlab auth issue (#79)
* Fix the gitlab oauth issue.

* Update for gitlab 11.1+

Versions beyond 11.1 (and possibly a few releases before) use a
different method for delivering tokens.  They also have disabled
version 3 of the api.

These changes address that and add a debugging mode for the server
that make it easier to debug issues like this in the future.

* Cleanup of PR.

Updated README. Removed code duplication.